fix: generate suggested locators upon selecting element instead of upon retrieving source (#1389)
* add slight caching for retrieved locator map

* remove unused third param for getOptimalXPath calls

* add comment to point to next actions

* WIP moving xpath calc to getLocators

* chore: remove unnecessary filter call

* chore: move findElementByPath to utils

* chore: remove unused state vars

* chore: revert to object for getLocators response

* chore: fix getLocators parameter

* chore: move getLocators to utils file

* start work on the complex locator detection

* chore: rename various source variables for clarity

* chore: move DOM conversion from areAttrAndValueUnique

* chore: add function to retrieve DOM node using path

* chore: update docstrings

* chore: save DOMParser in a const

* fix: calculate optimal xpath only on demand

* chore: do not pass local consts as parameters

* fix: enable predicate string for XCUIElementTypeApplication

* chore: adjust comments for getOptimalClassChain

* fix: calculate optimal class chain/predicate only on demand

* chore: define NATIVE_APP only once

* chore: update more var names

* chore: remove dedicated isIOS variable

* fix: skip accessibility id strategy in webviews

* chore: simplify childNodesOf

* chore: lint & format

* test: update unit tests

* test: add tests for accessibility id and class name

* test: add tests for path lookup methods

* test: add tests for class chain/predicate + format

* chore: split utils file

* chore: address review comments

* chore: update typedefs

* chore: update typedef again
