+++ app/renderer/components/Inspector/SelectedElement.js
@@ -13,9 +13,8 @@ import React, {useRef} from 'react';
 import {clipboard, shell} from '../../polyfills';
 import {ALERT, ROW} from '../AntdTypes';
 import styles from './Inspector.css';
-import {getLocators} from './shared';
+import {NATIVE_APP} from './shared';
 
-const NATIVE_APP = 'NATIVE_APP';
 const CLASS_CHAIN_DOCS_URL =
   'https://github.com/facebookarchive/WebDriverAgent/wiki/Class-Chain-Queries-Construction-Rules';
 const PREDICATE_DOCS_URL =
@@ -34,7 +33,6 @@ const SelectedElement = (props) => {
     isFindingElementsTimes,
     selectedElement,
     selectedElementId,
-    sourceXML,
     elementInteractionsNotAvailable,
     selectedElementSearchInProgress,
     t,
@@ -42,7 +40,6 @@ const SelectedElement = (props) => {
 
   const sendKeys = useRef();
 
-  const {attributes, classChain, predicateString, xpath} = selectedElement;
   const isDisabled = selectedElementSearchInProgress || isFindingElementsTimes;
 
   const selectedElementTableCell = (text, copyToClipBoard) => {
@@ -60,6 +57,15 @@ const SelectedElement = (props) => {
       return <div className={styles['selected-element-table-cells']}>{text}</div>;
     }
   };
+
+  const locatorStrategyDocsLink = (name, docsLink) => (
+    <span>
+      {name}
+      <strong>
+        <a onClick={(e) => e.preventDefault() || shell.openExternal(docsLink)}>&nbsp;(docs)</a>
+      </strong>
+    </span>
+  );
 
   // Get the columns for the attributes table
   let attributeColumns = [
@@ -79,8 +85,7 @@ const SelectedElement = (props) => {
   ];
 
   // Get the data for the attributes table
-  let attrArray = _.toPairs(attributes).filter(([key]) => key !== 'path');
-  let dataSource = attrArray.map(([key, value]) => ({
+  let dataSource = _.toPairs(selectedElement.attributes).map(([key, value]) => ({
     key,
     value,
     name: key,
@@ -116,66 +121,26 @@ const SelectedElement = (props) => {
   }
 
   // Get the data for the strategies table
-  let findDataSource = _.toPairs(getLocators(attributes, sourceXML)).map(([key, selector]) => ({
+  let findDataSource = selectedElement.strategyMap.map(([key, selector]) => ({
     key,
     selector,
     find: key,
   }));
 
-  // If XPath is the only provided data source, warn the user about it's brittleness
-  let showXpathWarning = false;
-  if (findDataSource.length === 0) {
-    showXpathWarning = true;
+  // Add documentation links to supported strategies
+  for (const locator of findDataSource) {
+    switch (locator.key) {
+      case '-ios class chain':
+        locator.find = locatorStrategyDocsLink(locator.key, CLASS_CHAIN_DOCS_URL);
+        break;
+      case '-ios predicate string':
+        locator.find = locatorStrategyDocsLink(locator.key, PREDICATE_DOCS_URL);
+        break;
+    }
   }
 
-  // Add class chain to the data source as well
-  if (classChain && currentContext === NATIVE_APP) {
-    const classChainText = (
-      <span>
-        -ios class chain
-        <strong>
-          <a onClick={(e) => e.preventDefault() || shell.openExternal(CLASS_CHAIN_DOCS_URL)}>
-            &nbsp;(docs)
-          </a>
-        </strong>
-      </span>
-    );
-
-    findDataSource.push({
-      key: '-ios class chain',
-      find: classChainText,
-      selector: classChain,
-    });
-  }
-
-  // Add predicate string to the data source as well
-  if (predicateString && currentContext === NATIVE_APP) {
-    const predicateStringText = (
-      <span>
-        -ios predicate string
-        <strong>
-          <a onClick={(e) => e.preventDefault() || shell.openExternal(PREDICATE_DOCS_URL)}>
-            &nbsp;(docs)
-          </a>
-        </strong>
-      </span>
-    );
-
-    findDataSource.push({
-      key: '-ios predicate string',
-      find: predicateStringText,
-      selector: predicateString,
-    });
-  }
-
-  // Add XPath to the data source as well
-  if (xpath) {
-    findDataSource.push({
-      key: 'xpath',
-      find: 'xpath',
-      selector: xpath,
-    });
-  }
+  // If XPath is the only optimal selector, warn the user about its brittleness
+  const showXpathWarning = findDataSource.length === 1;
 
   // Replace table data with table data that has the times
   if (findElementsExecutionTimes.length > 0) {
