+++ app/renderer/containers/InspectorPage.js
@@ -2,7 +2,7 @@ import {connect} from 'react-redux';
 
 import * as InspectorActions from '../actions/Inspector';
 import InspectorPage from '../components/Inspector/Inspector';
-import {withTranslation} from '../util';
+import {withTranslation} from '../utils/other';
 
 function mapStateToProps(state) {
   return state.inspector;
